Move RequestPanel into its own folder
