+++ frontend/src/pages/RequestorPage/RequestPanel/index.ts
@@ -1,0 +1,1 @@
+export { RequestPanel } from "./RequestPanel";
